Ui on process

+++ src/DreamGallery.Domain/Entities/Purchase.cs
@@ -4,8 +4,8 @@ namespace DreamGallery.Domain.Entities
 {
     public class Purchase : Auditable
     {
-        public int UserId { get; set; }
-        public int ArtworkId { get; set; }
+        public long UserId { get; set; }
+        public long ArtworkId { get; set; }
         public DateTime PurchaseDate { get; set; }
     }
 }
